extract common utils for prototype transform tests
